SD-113: rename integ test classes
