+++ src/apiTest/kotlin/com/ampnet/auditornode/script/api/objects/ConvertersApiTest.kt
@@ -1,26 +1,23 @@
 package com.ampnet.auditornode.script.api.objects
 
 import assertk.assertThat
-import com.ampnet.auditornode.TestBase
-import com.ampnet.auditornode.isRightContaining
+import com.ampnet.auditornode.ApiTestBase
+import com.ampnet.auditornode.isJsonEqualTo
 import com.ampnet.auditornode.jsAssertions
-import com.ampnet.auditornode.script.api.ExecutionContext
-import com.ampnet.auditornode.script.api.classes.HttpClient
+import com.ampnet.auditornode.model.response.ExecuteScriptOkResponse
 import com.ampnet.auditornode.script.api.model.SuccessfulAudit
-import com.ampnet.auditornode.service.impl.JavaScriptAuditingService
+import io.micronaut.http.HttpRequest
+import io.micronaut.http.MediaType
+import io.micronaut.test.extensions.junit5.annotation.MicronautTest
 import org.intellij.lang.annotations.Language
 import org.junit.jupiter.api.Test
-import org.mockito.kotlin.mock
 
-class ConvertersJavaScriptTest : TestBase() {
-
-    private val httpClient = HttpClient(mock())
-    private val environment = Properties(mock())
-    private val service = JavaScriptAuditingService(httpClient, environment)
+@MicronautTest
+class ConvertersApiTest : ApiTestBase() {
 
     @Test
-    fun `must correctly convert JS array into a list and vice-versa`() {
-        verify("JS array is correctly converted into a list and then back into an array") {
+    fun `must correctly execute auditing script which uses Converters list API`() {
+        verify("list converters work correctly") {
             @Language("JavaScript") val scriptSource = jsAssertions + """
                 function audit(auditData) {
                     const array = [1, 2, "three"];
@@ -43,14 +40,20 @@ class ConvertersJavaScriptTest : TestBase() {
                     return AuditResult.success();
                 }
             """.trimIndent()
-            val result = service.evaluate(scriptSource, ExecutionContext.noOp)
-            assertThat(result).isRightContaining(SuccessfulAudit)
+
+            val result = client.toBlocking().retrieve(
+                HttpRequest.POST("/script/execute", scriptSource).apply {
+                    contentType(MediaType.TEXT_PLAIN_TYPE)
+                }
+            )
+
+            assertThat(result).isJsonEqualTo(ExecuteScriptOkResponse(SuccessfulAudit))
         }
     }
 
     @Test
-    fun `must correctly convert JS object into a map and vice-versa`() {
-        verify("JS object is correctly converted into a map and then back into an object") {
+    fun `must correctly execute auditing script which uses Converters map API`() {
+        verify("map converters work correctly") {
             @Suppress("JSUnfilteredForInLoop")
             @Language("JavaScript") val scriptSource = jsAssertions + """
                 function audit(auditData) {
@@ -87,8 +90,14 @@ class ConvertersJavaScriptTest : TestBase() {
                     return AuditResult.success();
                 }
             """.trimIndent()
-            val result = service.evaluate(scriptSource, ExecutionContext.noOp)
-            assertThat(result).isRightContaining(SuccessfulAudit)
+
+            val result = client.toBlocking().retrieve(
+                HttpRequest.POST("/script/execute", scriptSource).apply {
+                    contentType(MediaType.TEXT_PLAIN_TYPE)
+                }
+            )
+
+            assertThat(result).isJsonEqualTo(ExecuteScriptOkResponse(SuccessfulAudit))
         }
     }
 }
