+++ src/apiTest/kotlin/com/ampnet/auditornode/script/api/objects/PropertiesApiTest.kt
@@ -13,7 +13,7 @@ import org.intellij.lang.annotations.Language
 import org.junit.jupiter.api.Test
 
 @MicronautTest(propertySources = ["script-properties-test.yaml"])
-class PropertiesJavaScriptApiTest : ApiTestWithPropertiesBase("script-properties-test") {
+class PropertiesApiTest : ApiTestWithPropertiesBase("script-properties-test") {
 
     @Test
     fun `must correctly execute auditing script which uses script properties`() {
