+++ src/integTest/kotlin/com/ampnet/auditornode/script/api/classes/NoOpIpfsIntegTest.kt
@@ -12,7 +12,7 @@ import org.intellij.lang.annotations.Language
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.mock
 
-class NoOpIpfsTest : TestBase() {
+class NoOpIpfsIntegTest : TestBase() {
 
     private val httpClient = HttpClient(mock())
     private val environment = Properties(mock())
