+++ src/integTest/kotlin/com/ampnet/auditornode/script/api/classes/WebSocketOutputIntegTest.kt
@@ -23,7 +23,7 @@ import org.mockito.kotlin.reset
 import org.mockito.kotlin.then
 import org.mockito.kotlin.times
 
-class WebSocketOutputTest : TestBase() {
+class WebSocketOutputIntegTest : TestBase() {
 
     private val session = mock<WebSocketSession>()
     private val webSocketApi = WebSocketApi(session, TestUtils.objectMapper)
