+++ src/integTest/kotlin/com/ampnet/auditornode/script/api/objects/PropertiesIntegTest.kt
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test
 import org.mockito.kotlin.given
 import org.mockito.kotlin.mock
 
-class PropertiesJavaScriptTest : TestBase() {
+class PropertiesIntegTest : TestBase() {
 
     private val httpClient = HttpClient(mock())
